refactored edit an organisation component and remove unused JSON data at organisations folder
