--- frontend/src/Components/Organisation/org_data.js
@@ -1,32 +1,0 @@
-const orgData = [
-  {
-    id: 1,
-    name: 'Haringey Service',
-    area: 'North',
-    borough: 'Haringey',
-    advice: 'Help accessing NHS',
-  },
-  {
-    id: 2,
-    name: 'Haringey Service',
-    area: 'North',
-    borough: 'Haringey',
-    advice: 'Help accessing NHS',
-  },
-  {
-    id: 3,
-    name: 'Haringey Service',
-    area: 'North',
-    borough: 'Haringey',
-    advice: 'Help accessing NHS',
-  },
-  {
-    id: 4,
-    name: 'Haringey Service',
-    area: 'North',
-    borough: 'Haringey',
-    advice: 'Help accessing NHS',
-  },
-];
-
-export default orgData;
