--- frontend/src/Components/Organisation/org_data.json
@@ -1,30 +1,0 @@
-[
-  {
-    "id": 1,
-    "name": "Haringey Service",
-    "area": "North",
-    "borough": "Haringey",
-    "advice": "Help accessing NHS"
-  },
-  {
-    "id": 2,
-    "name": "Haringey Service",
-    "area": "North",
-    "borough": "Haringey",
-    "advice": "Help accessing NHS"
-  },
-  {
-    "id": 3,
-    "name": "Haringey Service",
-    "area": "North",
-    "borough": "Haringey",
-    "advice": "Help accessing NHS"
-  },
-  {
-    "id": 4,
-    "name": "Haringey Service",
-    "area": "North",
-    "borough": "Haringey",
-    "advice": "Help accessing NHS"
-  }
-]
