+++ frontend/src/Components/TopNav/index.js
@@ -82,7 +82,7 @@ class TopNav extends Component {
                 <Link to={`/${titleLink}`}>
                   {category}
                 </Link>
-                {addOrg || homePage || role !== 'Admin' && role !== 'Editor' ? null :
+                {addOrg || homePage || (role !== 'Admin' && role !== 'Editor') ? null :
                   (
                     <Link to={`/${addLink}`} className="add-orgnaization">
                       <Button
